+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.2.1] - 2020-11-06
+
 ## [1.2.0] - 2020-10-20
 
 ## [1.1.0] - 2020-10-06